was failing to copy empty directories

+++ scripts/labtainer-student/bin/CheckTars.py
@@ -68,17 +68,27 @@ def newest_file_in_tree(rootfolder):
 
 def copydir(source, dest):
     """Copy a directory structure overwriting existing files"""
+    dest_par = os.path.dirname(dest)
     for root, dirs, files in os.walk(source):
         if not os.path.isdir(root):
             os.makedirs(root)
 
+        for mdir in dirs:
+            try:
+                dest_path = os.path.join(dest_par, root, mdir)
+                if not os.path.isdir(dest_path):
+                    os.makedirs(dest_path)
+            except:
+                pass
         for file in files:
             rel_path = root.replace(source, '').lstrip(os.sep)
             dest_path = os.path.join(dest, rel_path)
-
             if not os.path.isdir(dest_path):
                 os.makedirs(dest_path)
-            shutil.copyfile(os.path.join(root, file), os.path.join(dest_path, file))
+            cpy_src = os.path.join(root, file)
+            cpy_dest = os.path.join(dest_path, file)
+            shutil.copyfile(cpy_src, cpy_dest)
+            shutil.copymode(cpy_src, cpy_dest)
 
 def CheckTars(container_dir, image_name, logger):
     here = os.getcwd()
@@ -92,11 +102,9 @@ def CheckTars(container_dir, image_name, logger):
     for f in tar_list:
         full = os.path.join(container_dir, f)
         if os.path.isdir(full) and f.endswith('_tar'):
-            try:
+            if os.path.isdir(tmp_loc):
                 logger.debug('remove tree at %s' % tmp_loc)
                 shutil.rmtree(tmp_loc)
-            except:
-                pass
             os.mkdir(tmp_loc)
             os.chdir(full)
             tmp_name = f[:-4]
@@ -121,8 +129,10 @@ def CheckTars(container_dir, image_name, logger):
                         logger.debug('cfile is %s' % cfile)
                         if cfile != external:
                             if os.path.isdir(cfile):
+                                logger.debug('copydir %s' % cfile)
                                 copydir(cfile, os.path.join(tmp_loc, cfile))
                             else:
+                                logger.debug('copytree %s' % cfile)
                                 shutil.copytree(cfile, os.path.join(tmp_loc, cfile))
                     os.chdir(tmp_loc)
                     full_tar = os.path.join(full, tar_name)
@@ -131,7 +141,7 @@ def CheckTars(container_dir, image_name, logger):
                     else:
                         cmd = 'tar czf %s `ls -A -1`' % (full_tar)
                     os.system(cmd)
-                    #print('did %s' % cmd)
+                    logger.debug('did %s' % cmd)
             else:
                 ''' is a tar file, should it be updated? '''
                 os.chdir(full)
